business logic redefined

+++ Api/Services/ContactService.cs
@@ -18,28 +18,28 @@ public class ContactService : IContactService
         _mapper = mapper;
     }
 
-    public async Task<List<ContactResponse>> GetAll()
+    public async Task<List<GetAllContactResponse>> GetAll()
     {
         var contacts = await _contactRepository.GetAll();
-        return _mapper.Map<List<ContactResponse>>(contacts);
+        return _mapper.Map<List<GetAllContactResponse>>(contacts);
     }
 
-    public async Task<ContactResponse> GetById(int id)
+    public async Task<GetByIdContactResponse> GetById(int id)
     {
         var contact = await _contactRepository.GetById(id);
         if (contact is null)
         {
             throw new ContactException("Can't GetById: contact with given Id don't exist");
         }
-        return _mapper.Map<ContactResponse>(contact);
+        return _mapper.Map<GetByIdContactResponse>(contact);
     }
 
-    public async Task<ContactResponse> Create(AddContactRequest item)
+    public async Task<CreateContactResponse> Create(CreateContactRequest create)
     {
-        AddContactRequestValidator validator = new();
-        validator.ValidateAndThrow(item);
+        CreateContactRequestValidator validator = new();
+        validator.ValidateAndThrow(create);
 
-        var contact = _mapper.Map<Contact>(item);
+        var contact = _mapper.Map<Contact>(create);
 
         var exists = await _contactRepository.GetById(contact.Id);
         if (exists is not null) 
@@ -49,30 +49,42 @@ public class ContactService : IContactService
 
         await _contactRepository.Create(contact);
 
-        return _mapper.Map<ContactResponse>(contact);
+        return _mapper.Map<CreateContactResponse>(contact);
     }
 
-    public async Task<ContactResponse> Update(AddContactRequest item)
+    public async Task<List<FindContactResponse>> Find(FindContactRequest find)
     {
-        AddContactRequestValidator validator = new();
-        validator.ValidateAndThrow(item);
+        FindContactRequestValidator validator = new();
+        validator.ValidateAndThrow(find);
 
-        var contact = _mapper.Map<Contact>(item);
+        var found = await _contactRepository.Find(find.Keyword);
 
-        var exists = await _contactRepository.GetById(contact.Id) ??
-            throw new ContactException("Can't Update: contact with given Id don't exist"); 
-
-        await _contactRepository.Update(exists, contact);
-
-        return _mapper.Map<ContactResponse>(contact);
+        return _mapper.Map<List<FindContactResponse>>(found);
     }
 
-    public async Task<ContactResponse> DeleteById(int id)
+    public async Task<UpdateContactResponse> Update(UpdateContactRequest update)
     {
+        UpdateContactRequestValidator validator = new();
+        validator.ValidateAndThrow(update);
+
+        var contact = _mapper.Map<Contact>(update);
+
+        _ = await _contactRepository.GetById(contact.Id) ??
+            throw new ContactException("Can't Update: contact with given Id don't exist");
+
+        await _contactRepository.Update(contact);
+
+        return _mapper.Map<UpdateContactResponse>(contact);
+    }
+
+    public async Task<DeleteContactResponse> DeleteById(int id)
+    {
+        if (id < 1) throw new ContactException($"Can't Delete: id = {id} isn't possible, id must be positive integer");
+
         var item = await _contactRepository.GetById(id) ?? 
             throw new ContactException("Can't Delete: contact with given Id don't exist");
         await _contactRepository.Delete(item);
 
-        return _mapper.Map<ContactResponse>(item);
+        return _mapper.Map<DeleteContactResponse>(item);
     }
 }
